separate out some traits
